+++ modules/ecs_service_fargate_elb/variables.tf
@@ -88,10 +88,10 @@ variable "target_group_arn" {
   description = "Target of an Amazon ALB (Load Balancer), as given by a load balancer."
 }
 
-variable "datadog_api_key" {
-  description = "Datadog API key (DD_API_KEY). This will activate the Datadog agent sidecar/replica container on the same ECS task. Please adjust var.cpu and/or var.memory to accommodate if necessary."
-  default     = ""
-  type        = string
+variable "datadog_enabled" {
+  description = "This option will activate the Datadog agent sidecar/replica container on the same ECS task. Please adjust var.cpu and/or var.memory to accommodate if necessary."
+  default     = false
+  type        = bool
 }
 
 variable "datadog_agent_version" {